Front Page added

+++ routes/web.php
@@ -23,7 +23,8 @@ use Illuminate\Support\Facades\Route;
 
 
 
-Route::get('/', [PlayerHistory::class, 'index'])->name('homepage.index');
+Route::get('/', [PlayerHistory::class, 'index_front'])->name('homepage.front');
+
 Route::get('results', [TournamentController::class, 'results'])->name('tournament.results');
 
 Route::group(['middleware' => ['auth']], function () {
@@ -36,6 +37,7 @@ Route::group(['middleware' => ['auth']], function () {
 
     });
 
+//    Route::get('/home', [PlayerHistory::class, 'index'])->name('homepage.index');
     Route::get('dashboard', [DashboardController::class, 'index'])->name('dashboard.index');
 
     Route::impersonate();
@@ -54,6 +56,11 @@ Route::group(['middleware' => ['auth']], function () {
         Route::post('password/{user}', [UsersController::class, 'password_update'])->name('user.password_update');
 
         Route::get('/reset', function () {
+
+            if (env('APP_ENV') != 'local') {
+                dd('Nothing to do');
+                return;
+            }
             \Artisan::call('migrate:fresh');
             \Artisan::call('db:seed');
             \Artisan::call('optimize:clear');
@@ -61,6 +68,5 @@ Route::group(['middleware' => ['auth']], function () {
         });
 
     });
-
 
 });
